Moving enums to different folder. Exporting interface in index file.
