+++ src/enums/affiliationInvitation.ts
@@ -1,0 +1,10 @@
+export enum AffiliationInvitationType {
+  EMAIL = 'EMAIL',
+  REQUEST = 'REQUEST',
+}
+
+export enum AffiliationInvitationStatus {
+  PENDING = 'PENDING',
+  FAILED = 'FAILED',
+  ACCEPTED = 'ACCEPTED',
+}
