+++ src/interfaces/affiliation-invitation-interface.ts
@@ -1,13 +1,4 @@
-export enum AffiliationInvitationType {
-  EMAIL = 'EMAIL',
-  REQUEST = 'REQUEST',
-}
-
-export enum AffiliationInvitationStatus {
-  PENDING = 'PENDING',
-  FAILED = 'FAILED',
-  ACCEPTED = 'ACCEPTED',
-}
+import { AffiliationInvitationStatus, AffiliationInvitationType } from '@/enums'
 
 export interface AffiliationInvitationIF {
   id: number
