+++ src/interfaces/index.ts
@@ -14,6 +14,7 @@ export type { AlterationIF, ContactPointIF, NameTranslationIF, ShareStructureIF,
 // must come before other interfaces that use it
 export * from './officer-interface'
 export * from './address-interfaces'
+export * from './affiliation-invitation-interface'
 export * from './alert-message-interface'
 export * from './api-filing-interface'
 export * from './api-task-interface'
